refactor: add component CommentsSortingPopover
